sublcass formatting
