+++ .gitignore
@@ -12,3 +12,4 @@ uncertainties.egg-info/
 .cache
 # vim temporary files
 .*.swp
+.coverage
