+++ uncertainties/core/core.py
@@ -44,8 +44,7 @@ from . compat import (deprecation, FLOAT_LIKE_TYPES, CONSTANT_TYPES, CallableStd
 from . util import (set_doc, covariance_matrix, partial_derivative,
 NumericalDerivatives, IndexableIter, basestring,
 )
-from . formatting import (first_digit, PDG_precision, format_num, signif_dgt_to_limit
-)
+from . formatting import UFloatFormatting
 from . parsing import (str_to_number_with_uncert,)
 
 try:
@@ -624,7 +623,7 @@ class LinearCombination(object):
     def __setstate__(self, state):
         (self.linear_combo,) = state
 
-class AffineScalarFunc(object):
+class AffineScalarFunc(UFloatFormatting):
     """
     Affine functions that support basic mathematical operations
     (addition, etc.).  Such functions can for instance be used for
@@ -902,490 +901,6 @@ class AffineScalarFunc(object):
         # string
         # (http://docs.python.org/2/library/string.html#format-specification-mini-language):
         return self.format('')
-
-    def __format__(self, format_spec):
-        '''
-        Formats a number with uncertainty.
-
-        The format specification are the same as for format() for
-        floats, as defined for Python 2.6+ (restricted to what the %
-        operator accepts, if using an earlier version of Python),
-        except that the n presentation type is not supported. In
-        particular, the usual precision, alignment, sign flag,
-        etc. can be used. The behavior of the various presentation
-        types (e, f, g, none, etc.) is similar. Moreover, the format
-        is extended: the number of digits of the uncertainty can be
-        controlled, as is the way the uncertainty is indicated (with
-        +/- or with the short-hand notation 3.14(1), in LaTeX or with
-        a simple text string,...).
-
-        Beyond the use of options at the end of the format
-        specification, the main difference with floats is that a "u"
-        just before the presentation type (f, e, g, none, etc.)
-        activates the "uncertainty control" mode (e.g.: ".6u").  This
-        mode is also activated when not using any explicit precision
-        (e.g.: "g", "10f", "+010,e" format specifications).  If the
-        uncertainty does not have a meaningful number of significant
-        digits (0 and NaN uncertainties), this mode is automatically
-        deactivated.
-
-        The nominal value and the uncertainty always use the same
-        precision. This implies trailing zeros, in general, even with
-        the g format type (contrary to the float case). However, when
-        the number of significant digits of the uncertainty is not
-        defined (zero or NaN uncertainty), it has no precision, so
-        there is no matching. In this case, the original format
-        specification is used for the nominal value (any "u" is
-        ignored).
-
-        Any precision (".p", where p is a number) is interpreted (if
-        meaningful), in the uncertainty control mode, as indicating
-        the number p of significant digits of the displayed
-        uncertainty. Example: .1uf will return a string with one
-        significant digit in the uncertainty (and no exponent).
-
-        If no precision is given, the rounding rules from the
-        Particle Data Group are used, if possible
-        (http://pdg.lbl.gov/2010/reviews/rpp2010-rev-rpp-intro.pdf). For
-        example, the "f" format specification generally does not use
-        the default 6 digits after the decimal point, but applies the
-        PDG rules.
-
-        A common exponent is used if an exponent is needed for the
-        larger of the nominal value (in absolute value) and the
-        standard deviation, unless this would result in a zero
-        uncertainty being represented as 0e... or a NaN uncertainty as
-        NaNe.... Thanks to this common exponent, the quantity that
-        best describes the associated probability distribution has a
-        mantissa in the usual 1-10 range. The common exponent is
-        factored (as in "(1.2+/-0.1)e-5"). unless the format
-        specification contains an explicit width (" 1.2e-5+/- 0.1e-5")
-        (this allows numbers to be in a single column, when printing
-        numbers over many lines). Specifying a minimum width of 1 is a
-        way of forcing any common exponent to not be factored out.
-
-        The fill, align, zero and width parameters of the format
-        specification are applied individually to each of the nominal
-        value and standard deviation or, if the shorthand notation is
-        used, globally.
-
-        The sign parameter of the format specification is only applied
-        to the nominal value (since the standard deviation is
-        positive).
-
-        In the case of a non-LaTeX output, the returned string can
-        normally be parsed back with ufloat_fromstr(). This however
-        excludes cases where numbers use the "," thousands separator,
-        for example.
-
-        Options can be added, at the end of the format
-        specification. Multiple options can be specified:
-
-        - When "P" is present, the pretty-printing mode is activated: "±"
-          separates the nominal value from the standard deviation, exponents
-          use superscript characters, etc.
-        - When "S" is present (like in .1uS), the short-hand notation 1.234(5)
-          is used, indicating an uncertainty on the last digits; if the digits
-          of the uncertainty straddle the decimal point, it uses a fixed-point
-          notation, like in 12.3(4.5).
-        - When "L" is present, the output is formatted with LaTeX.
-        - "p" ensures that there are parentheses around the …±… part (no
-          parentheses are added if some are already present, for instance
-          because of an exponent or of a trailing % sign, etc.). This produces
-          outputs like (1.0±0.2) or (1.0±0.2)e7, which can be useful for
-          removing any ambiguity if physical units are added after the printed
-          number.
-    
-        An uncertainty which is exactly zero is represented as the
-        integer 0 (i.e. with no decimal point).
-
-        The "%" format type forces the percent sign to be at the end
-        of the returned string (it is not attached to each of the
-        nominal value and the standard deviation).
-
-        Some details of the formatting can be customized as described
-        in format_num().
-        '''
-
-        # Convention on limits "between" digits: 0 = exactly at the
-        # decimal point, -1 = after the first decimal, 1 = before the
-        # units digit, etc.
-
-        # Convention on digits: 0 is units (10**0), 1 is tens, -1 is
-        # tenths, etc.
-
-        # This method does the format specification parsing, and
-        # calculates the various parts of the displayed value
-        # (mantissas, exponent, position of the last digit). The
-        # formatting itself is delegated to format_num().
-
-        ########################################
-
-        # Format specification parsing:
-
-        match = re.match(r'''
-            (?P<fill>[^{}]??)(?P<align>[<>=^]?)  # fill cannot be { or }
-            (?P<sign>[-+ ]?)
-            (?P<zero>0?)
-            (?P<width>\d*)
-            (?P<comma>,?)
-            (?:\.(?P<prec>\d+))?
-            (?P<uncert_prec>u?)  # Precision for the uncertainty?
-            # The type can be omitted. Options must not go here:
-            (?P<type>[eEfFgG%]??)  # n not supported
-            (?P<options>[PSLp]*)  # uncertainties-specific flags
-            $''',
-            format_spec,
-            re.VERBOSE)
-
-        # Does the format specification look correct?
-        if not match:
-            raise ValueError(
-                'Format specification %r cannot be used with object of type'
-                ' %r. Note that uncertainties-specific flags must be put at'
-                ' the end of the format string.'
-                # Sub-classes handled:
-                % (format_spec, self.__class__.__name__))
-
-        # Effective format presentation type: f, e, g, etc., or None,
-        # like in
-        # https://docs.python.org/3.4/library/string.html#format-specification-mini-language. Contrary
-        # to what is written in the documentation, it is not true that
-        # None is "the same as 'g'": "{}".format() and "{:g}" do not
-        # give the same result, on 31415000000.0. None is thus kept as
-        # is instead of being replaced by "g".
-        pres_type = match.group('type') or None
-
-        # Shortcut:
-        fmt_prec = match.group('prec')  # Can be None
-
-        ########################################
-
-        # Since the '%' (percentage) format specification can change
-        # the value to be displayed, this value must first be
-        # calculated. Calculating the standard deviation is also an
-        # optimization: the standard deviation is generally
-        # calculated: it is calculated only once, here:
-        nom_val = self.nominal_value
-        std_dev = self.std_dev
-
-        # 'options' is the options that must be given to format_num():
-        options = set(match.group('options'))
-
-        ########################################
-
-        # The '%' format is treated internally as a display option: it
-        # should not be applied individually to each part:
-        if pres_type == '%':
-            # Because '%' does 0.0055*100, the value
-            # 0.5499999999999999 is obtained, which rounds to 0.5. The
-            # original rounded value is 0.006. The same behavior is
-            # found in Python 2.7: '{:.1%}'.format(0.0055) is '0.5%'.
-            # If a different behavior is needed, a solution to this
-            # problem would be to do the rounding before the
-            # multiplication.
-            std_dev *= 100
-            nom_val *= 100
-            pres_type = 'f'
-            options.add('%')
-
-        # At this point, pres_type is in eEfFgG or None (not %).
-
-        ########################################
-
-        # Non-real values (nominal value or standard deviation) must
-        # be handled in a specific way:
-        real_values = [value for value in [abs(nom_val), std_dev]
-                       if not isinfinite(value)]
-
-        # Calculation of digits_limit, which defines the precision of
-        # the nominal value and of the standard deviation (it can be
-        # None when it does not matter, like for NaN±NaN):
-
-        # Reference value for the calculation of a possible exponent,
-        # if needed:
-        if pres_type in (None, 'e', 'E', 'g', 'G'):
-            # Reference value for the exponent: the largest value
-            # defines what the exponent will be (another convention
-            # could have been chosen, like using the exponent of the
-            # nominal value, irrespective of the standard deviation):
-            try:
-                exp_ref_value = max(real_values)
-            except ValueError:  # No non-NaN value: NaN±NaN…
-                # No meaningful common exponent can be obtained:
-                pass
-            ## else:
-            ##     print "EXP_REF_VAL", exp_ref_value
-
-        # Should the precision be interpreted like for a float, or
-        # should the number of significant digits on the uncertainty
-        # be controlled?
-        if ((
-            # Default behavior: number of significant digits on the
-            # uncertainty controlled (if useful, i.e. only in
-            # situations where the nominal value and the standard
-            # error digits are truncated at the same place):
-            (not fmt_prec and len(real_values)==2)
-             or match.group('uncert_prec'))  # Explicit control
-            # The number of significant digits of the uncertainty must
-            # be meaningful, otherwise the position of the significant
-            # digits of the uncertainty does not have a clear
-            # meaning. This gives us the *effective* uncertainty
-            # control mode:
-            and std_dev
-            and not isinfinite(std_dev)):
-
-            # The number of significant digits on the uncertainty is
-            # controlled.
-
-            # The limit digits_limit on the digits of nom_val and std_dev
-            # to be displayed is calculated. If the exponent notation is
-            # used, this limit is generally different from the finally
-            # displayed limit (e.g. 314.15+/-0.01 has digits_limit=-2, but
-            # will be displayed with an exponent as (3.1415+/-0.0001)e+02,
-            # which corresponds to 4 decimals after the decimal point, not
-            # 2).
-
-            # Number of significant digits to use:
-            if fmt_prec:
-                num_signif_d = int(fmt_prec)  # Can only be non-negative
-                if not num_signif_d:
-                    raise ValueError("The number of significant digits"
-                                     " on the uncertainty should be positive")
-            else:
-                (num_signif_d, std_dev) = PDG_precision(std_dev)
-
-            digits_limit = signif_dgt_to_limit(std_dev, num_signif_d)
-
-        else:
-
-            # No control of the number of significant digits on the
-            # uncertainty.
-
-            ## print "PRECISION NOT BASED ON UNCERTAINTY"
-
-            # The precision has the same meaning as for floats (it is
-            # not the uncertainty that defines the number of digits).
-
-            # The usual default precision is used (this is useful for
-            # 3.141592±NaN with an "f" format specification, for
-            # example):
-            #
-            # prec is the precision for the main parts of the final
-            # format (in the sense of float formatting):
-            #
-            # https://docs.python.org/3.4/library/string.html#format-specification-mini-language
-            if fmt_prec:
-                prec = int(fmt_prec)
-            elif pres_type is None:
-                prec = 12
-            else:
-                prec = 6
-
-            if pres_type in ('f', 'F'):
-
-                digits_limit = -prec
-
-            else:  # Format type in None, eEgG
-
-                # We first calculate the number of significant digits
-                # to be displayed (if possible):
-
-                if pres_type in ('e', 'E'):
-                    # The precision is the number of significant
-                    # digits required - 1 (because there is a single
-                    # digit before the decimal point, which is not
-                    # included in the definition of the precision with
-                    # the e/E format type):
-                    num_signif_digits = prec+1
-
-                else:  # Presentation type in None, g, G
-
-                    # Effective format specification precision: the rule
-                    # of
-                    # http://docs.python.org/2.7/library/string.html#format-specification-mini-language
-                    # is used:
-
-                    # The final number of significant digits to be
-                    # displayed is not necessarily obvious: trailing
-                    # zeros are removed (with the gG presentation
-                    # type), so num_signif_digits is the number of
-                    # significant digits if trailing zeros were not
-                    # removed. This quantity is relevant for the
-                    # rounding implied by the exponent test of the g/G
-                    # format:
-
-                    # 0 is interpreted like 1 (as with floats with a
-                    # gG presentation type):
-                    num_signif_digits = prec or 1
-
-                # The number of significant digits is important for
-                # example for determining the exponent:
-
-                ## print "NUM_SIGNIF_DIGITS", num_signif_digits
-
-                digits_limit = (
-                    signif_dgt_to_limit(exp_ref_value, num_signif_digits)
-                    if real_values
-                    else None)
-
-                ## print "DIGITS_LIMIT", digits_limit
-
-        #######################################
-
-        # Common exponent notation: should it be used? use_exp is set
-        # accordingly. If a common exponent should be used (use_exp is
-        # True), 'common_exp' is set to the exponent that should be
-        # used.
-
-        if pres_type in ('f', 'F'):
-            use_exp = False
-        elif pres_type in ('e', 'E'):
-            if not real_values:
-                use_exp = False
-            else:
-                use_exp = True
-                # !! This calculation might have been already done,
-                # for instance when using the .0e format:
-                # signif_dgt_to_limit() was called before, which
-                # prompted a similar calculation:
-                common_exp = first_digit(round(exp_ref_value, -digits_limit))
-
-        else:  # None, g, G
-
-            # The rules from
-            # https://docs.python.org/3.4/library/string.html#format-specification-mini-language
-            # are applied.
-
-            # Python's native formatting (whose result could be parsed
-            # in order to determine whether a common exponent should
-            # be used) is not used because there is shared information
-            # between the nominal value and the standard error (same
-            # last digit, common exponent) and extracting this
-            # information from Python would entail parsing its
-            # formatted string, which is in principle inefficient
-            # (internally, Python performs calculations that yield a
-            # string, and the string would be parsed back into
-            # separate parts and numbers, which is in principle
-            # unnecessary).
-
-            # Should the scientific notation be used? The same rule as
-            # for floats is used ("-4 <= exponent of rounded value <
-            # p"), on the nominal value.
-
-            if not real_values:
-                use_exp = False
-            else:
-                # Common exponent *if* used:
-                common_exp = first_digit(round(exp_ref_value, -digits_limit))
-
-                # print "COMMON EXP TEST VALUE", common_exp
-                # print "LIMIT EXP", common_exp-digits_limit+1
-                # print "WITH digits_limit", digits_limit
-
-                # The number of significant digits of the reference value
-                # rounded at digits_limit is exponent-digits_limit+1:
-                if -4 <= common_exp < common_exp-digits_limit+1:
-                    use_exp = False
-                else:
-                    use_exp = True
-
-        ########################################
-
-        # Calculation of signif_limit (position of the significant
-        # digits limit in the final fixed point representations; this
-        # is either a non-positive number, or None), of
-        # nom_val_mantissa ("mantissa" for the nominal value,
-        # i.e. value possibly corrected for a factorized exponent),
-        # and std_dev_mantissa (similarly for the standard
-        # deviation). common_exp is also set to None if no common
-        # exponent should be used.
-
-        if use_exp:
-
-            # Not 10.**(-common_exp), for limit values of common_exp:
-            factor = 10.**common_exp
-
-            nom_val_mantissa = nom_val/factor
-            std_dev_mantissa = std_dev/factor
-            # Limit for the last digit of the mantissas:
-            signif_limit = digits_limit - common_exp
-
-        else:  # No common exponent
-
-            common_exp = None
-
-            nom_val_mantissa = nom_val
-            std_dev_mantissa = std_dev
-            signif_limit = digits_limit
-
-        ## print "SIGNIF_LIMIT", signif_limit
-
-        ########################################
-
-        # Format of the main (i.e. with no exponent) parts (the None
-        # presentation type is similar to the g format type):
-
-        main_pres_type = 'fF'[(pres_type or 'g').isupper()]
-
-        # The precision of the main parts must be adjusted so as
-        # to take into account the special role of the decimal
-        # point:
-        if signif_limit is not None:  # If signif_limit is pertinent
-            # The decimal point location is always included in the
-            # printed digits (e.g., printing 3456 with only 2
-            # significant digits requires to print at least four
-            # digits, like in 3456 or 3500).
-            #
-            # The max() is important for example for
-            # 1234567.89123+/-12345.678 with the f format: in this
-            # case, signif_limit is +3 (2 significant digits necessary
-            # for the error, as per the PDG rules), but the (Python
-            # float formatting) precision to be used for the main
-            # parts is 0 (all digits must be shown).
-            #
-            # The 1 for the None pres_type represents "at least one
-            # digit past the decimal point" of Python
-            # (https://docs.python.org/3.4/library/string.html#format-specification-mini-language). This
-            # is only applied for null uncertainties.
-            prec = max(-signif_limit,
-                       1 if pres_type is None and not std_dev
-                       else 0)
-        ## print "PREC", prec
-
-        ########################################
-
-        # print (
-        #     "FORMAT_NUM parameters: nom_val_mantissa={},"
-        #     " std_dev_mantissa={}, common_exp={},"
-        #     " match.groupdict()={}, prec={}, main_pres_type={},"
-        #     " options={}".format(
-        #     nom_val_mantissa, std_dev_mantissa, common_exp,
-        #     match.groupdict(),
-        #     prec,
-        #     main_pres_type,
-        #     options))
-
-        # Final formatting:
-        return format_num(nom_val_mantissa, std_dev_mantissa, common_exp,
-                          match.groupdict(),
-                          prec=prec,
-                          main_pres_type=main_pres_type,
-                          options=options)
-
-    # Alternate name for __format__, for use with Python < 2.6 (and
-    # other Python versions if the user so chooses: this helps moving
-    # code from Python 2.6 to more recent versions):
-    @set_doc("""
-        Return the same result as self.__format__(format_spec), or
-        equivalently as the format(self, format_spec) of Python 2.6+.
-
-        This method is meant to be used for formatting numbers with
-        uncertainties in Python < 2.6, with '... %s ...' %
-        num.format('.2e').
-        """)
-    def format(*args, **kwargs):
-        return args[0].__format__(*args[1:], **kwargs)
 
     def std_score(self, value):
         """
